added tardis ticking
